Moved plan simplification to the server
